🚚  rename: 서비스와 비스니스내에 존재했던 index 파일 이름을 lotto 로 변경
